restructure the rest of the files
