+++ src/decima/data/write_hdf5.py
