+++ src/decima/utils/variant.py
